# src/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server News service",
        "title": "News API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/news",
    "paths": {
        "/delete_news/{idNews}": {
            "delete": {
                "description": "Update news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Update news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idNews",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad path || body param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "Noticia no encontrada",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/get_news": {
            "get": {
                "description": "Get news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get news",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Default 0",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Defaul false",
                        "name": "total",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Default global -\u003e Values: global || student",
                        "name": "type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.NewsMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad query param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "StatusServiceUnavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/get_single_news/{slug}": {
            "get": {
                "description": "Get a single news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get a single news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.SingleNewsMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "No tienes acceso a esta noticia",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "No pudimos encontrar la noticia...",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "410": {
                        "description": "Esta noticia ya no está disponible",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/like_news/{idNews}": {
            "post": {
                "description": "Toggle Like news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Like news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idNews",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad path param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "Noticia no encontrada",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/new_news": {
            "post": {
                "description": "New news",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "New news",
                "parameters": [
                    {
                        "type": "string",
                        "name": "body",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 500,
                        "minLength": 3,
                        "type": "string",
                        "name": "headline",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "binary",
                        "name": "img",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "maxLength": 100,
                        "minLength": 3,
                        "type": "string",
                        "name": "title",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.SingleNewsMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "El titulo de la noticia ya está en uso",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable - NATS || DB Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/update_news/{idNews}": {
            "put": {
                "description": "Update news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Update news",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MongoID",
                        "name": "idNews",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forms.UpdateNewsDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/res.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "body": {
                                            "$ref": "#/definitions/smaps.SingleNewsMap"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad path || body param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "404": {
                        "description": "Noticia no encontrada",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "forms.UpdateNewsDTO": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "headline": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3
                },
                "img": {
                    "type": "string",
                    "format": "binary"
                },
                "title": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "638660ca141aa4ee9faf07e8"
                },
                "first_lastname": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "Rojas"
                },
                "name": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "Karen"
                },
                "second_lastname": {
                    "type": "string",
                    "x-omitempty": true,
                    "example": "Valdes"
                }
            }
        },
        "res.Response": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "services.Image": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "638660ca141aa4ee9faf07e8"
                },
                "key": {
                    "type": "string",
                    "example": "$dsK2!1"
                },
                "url": {
                    "type": "string",
                    "example": "https://repository.com/file/$dsK2!1"
                }
            }
        },
        "services.NewsResponse": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string",
                    "example": "638660ca141aa4ee9faf07e8"
                },
                "author": {
                    "x-omitempty": true,
                    "$ref": "#/definitions/models.User"
                },
                "body": {
                    "type": "string",
                    "example": "This is a body..."
                },
                "headline": {
                    "type": "string",
                    "example": "Example..."
                },
                "image": {
                    "$ref": "#/definitions/services.Image"
                },
                "like": {
                    "type": "boolean"
                },
                "likes": {
                    "type": "integer",
                    "example": 10
                },
                "status": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string",
                    "example": "Title !!"
                },
                "type": {
                    "type": "string",
                    "example": "global"
                },
                "update_date": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "upload_date": {
                    "type": "string",
                    "example": "2022-09-21T20:10:23.309+00:00"
                },
                "url": {
                    "type": "string",
                    "example": "title"
                }
            }
        },
        "smaps.NewsMap": {
            "type": "object",
            "properties": {
                "news": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/services.NewsResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 15
                }
            }
        },
        "smaps.SingleNewsMap": {
            "type": "object",
            "properties": {
                "news": {
                    "$ref": "#/definitions/services.NewsResponse"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "BearerJWTToken in Authorization Header",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "tags": [
        {
            "description": "Service of news",
            "name": "news"
        }
    ]
}